Use simdjson for parsing of endpoint responses

+++ include/util/BatchHelper.h
@@ -19,17 +19,17 @@
 #ifndef BATCHHELPER_H
 #define BATCHHELPER_H
 
-#include "util/Types.h"
-
 #include <set>
 #include <functional>
+
+#include "util/Types.h"
 
 namespace olu::util {
 
     class BatchHelper {
     public :
         static void doInBatches(const std::set<id_t>& set,
-                         const long elementsPerBatch,
+                         const size_t elementsPerBatch,
                          const std::function<void(std::set<id_t>)> &func) {
             std::vector vector(set.begin(), set.end());
             std::vector<std::set<id_t> > vectorBatches;
